+++ core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # better-react
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - wy-helper@1.1.2
+
 ## 2.0.1
 
 ### Patch Changes